leecher has main function out of '__main__.py'
