+++ services/leecher/leecher/__init__.py
@@ -1,0 +1,6 @@
+from .listener import main
+
+
+__all__ = (
+    "main",
+)
